Factor common code from AIGopan and AIClassic into AISimple
The first attempt to do this was using the same ID field for both passes,
and LLVM was considering the passes as duplicates, and was running only
one of them. The way it is done now, all functions are factored in
AISimple, but the ID field is redefined in each class, and passed to the
constructor.

One can check that both passes are run:

ex$ ../src/analyzer -i bin/ex2.bc -c |& grep Starting
Starting analysis: PF+LW
Starting analysis: PF
Starting analysis: AIGopan
Starting analysis: AIClassic